chore: generate spotify client from generic authedRestClient
